Improve error handling in Download method
Add additional error handling to the Download method in DefaultStorageClient, so we do not ignore the error which is returned by DownloadFile(). Ignoring it led to exit 0 in case of downloading a non existent file.
Now it exits with exit code 1 if you download a file that does not exist and prints out the error.

Co-Authored-By: s4heid <sebastian.heid45@gmail.com>